Use uint to store colors (#25)
Pros:
- Don't have to write `unchecked((int)...)`, `uint argb = 0xFFFF_FFFF` is enough
- Easier to understand bitwise operations

Cons:
- Google uses ints, so it's harder to translate from the original implementation

+++ MaterialColorUtilities.Maui/MauiAppBuilderExtensions.cs
@@ -9,7 +9,7 @@ public static class MauiAppBuilderExtensions
         => builder.UseMaterialDynamicColors(_ => { });
 
     public static MauiAppBuilder UseMaterialDynamicColors(this MauiAppBuilder builder, uint fallbackSeed)
-        => builder.UseMaterialDynamicColors(opt => opt.FallbackSeed = (int)fallbackSeed);
+        => builder.UseMaterialDynamicColors(opt => opt.FallbackSeed = fallbackSeed);
 
     public static MauiAppBuilder UseMaterialDynamicColors(
         this MauiAppBuilder builder,
@@ -30,7 +30,7 @@ public static class MauiAppBuilderExtensions
         this MauiAppBuilder builder, uint fallbackSeed
     )
         where TDynamicColorService : class, IMauiInitializeService
-        => builder.UseMaterialDynamicColors<TDynamicColorService>(opt => opt.FallbackSeed = (int)fallbackSeed);
+        => builder.UseMaterialDynamicColors<TDynamicColorService>(opt => opt.FallbackSeed = fallbackSeed);
 
     public static MauiAppBuilder UseMaterialDynamicColors<
         [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors)]
